requirements.gpu.txt
python-dotenv==0.1.0
tweepy==3.8.0
jupyter==1.0.0
tensorflow-gpu==1.14.0
pandas==0.24.2
python-dotenv==0.1.0
nltk==3.4.3
scikit-learn==0.21.2
emoji==0.5.2
seaborn==0.9.0
tqdm==4.32.2
matplotlib==3.0.3
click==7.0
google-cloud-storage==1.20.0
pydot==1.4.1
gensim==3.8.0
